+++ packages/docs/components/DataBind/stories/immediate.js
@@ -1,0 +1,4 @@
+import { registerComponent } from '@studiometa/js-toolkit';
+import { DataBind } from '@studiometa/ui';
+
+registerComponent(DataBind);