+++ packages/docs/components/DataEffect/stories/effect-example.js
@@ -1,0 +1,6 @@
+import { registerComponent } from '@studiometa/js-toolkit';
+import { DataBind, DataEffect, DataModel } from '@studiometa/ui';
+
+registerComponent(DataBind);
+registerComponent(DataEffect);
+registerComponent(DataModel);