+++ packages/ui/Data/DataBind.ts
@@ -1,14 +1,12 @@
-import { Base } from '@studiometa/js-toolkit';
+import { Base, withGroup } from '@studiometa/js-toolkit';
 import type { BaseConfig, BaseProps } from '@studiometa/js-toolkit';
-import { isArray } from '@studiometa/js-toolkit/utils';
+import { isArray, nextTick } from '@studiometa/js-toolkit/utils';
 import { isInput, isCheckbox, isSelect } from './utils.js';
-
-const groups = new Map<string, Set<DataBind>>();
 
 export interface DataBindProps extends BaseProps {
   $options: {
     prop: string;
-    group: string;
+    immediate: boolean;
   };
 }
 
@@ -16,27 +14,20 @@ export interface DataBindProps extends BaseProps {
  * DataBind class.
  * @link https://ui.studiometa.dev/-/components/DataBind/
  */
-export class DataBind<T extends BaseProps = BaseProps> extends Base<DataBindProps & T> {
+export class DataBind<T extends BaseProps = BaseProps> extends withGroup(Base, 'data:')<DataBindProps & T> {
   static config: BaseConfig = {
     name: 'DataBind',
     options: {
       prop: String,
-      group: String,
+      immediate: Boolean,
     },
   };
 
+  /**
+   * @deprecated Use the `$group` getter instead.
+   */
   get relatedInstances() {
-    const { group } = this.$options;
-
-    const instances = groups.get(group) ?? groups.set(group, new Set()).get(group);
-
-    for (const instance of instances) {
-      if (!instance.$el.isConnected) {
-        instances.delete(instance);
-      }
-    }
-
-    return instances;
+    return this.$group as Set<this>;
   }
 
   get multiple() {
@@ -80,7 +71,7 @@ export class DataBind<T extends BaseProps = BaseProps> extends Base<DataBindProp
 
     if (isSelect(target)) {
       if (multiple) {
-        const values = [];
+        const values = [] as string[];
         // @ts-ignore
         for (const option of target.options) {
           if (option.selected) {
@@ -92,7 +83,7 @@ export class DataBind<T extends BaseProps = BaseProps> extends Base<DataBindProp
       }
 
       const option = target.options.item(target.selectedIndex);
-      return option.value;
+      return option?.value;
     }
 
     if (isCheckbox(target)) {
@@ -148,10 +139,10 @@ export class DataBind<T extends BaseProps = BaseProps> extends Base<DataBindProp
   }
 
   mounted() {
-    this.relatedInstances.add(this);
-  }
-
-  destroyed() {
-    this.relatedInstances.delete(this);
+    if (this.$options.immediate) {
+      nextTick().then(() => {
+        this.set(this.get());
+      });
+    }
   }
 }